Compare against last successful build if upstream matches head
This addresses the problem where a push to master would re-run
everything regardless of changes. This is not desired. Now, compares
against the commit of the last successful build of the upstream branch
(master, in most cases).

Force pushes to master should (can't test until this is committed)
compare against the last successful master build, which should be the
build before master was force-pushed.