Test view selectors only need to be functions on the first test view
It's inconvenient to need to wrap all the selectors in a function,
especially when you aren't parameterizing them. This change makes it so
you don't need to do that anymore. Also, this makes it more explicit
that parameterized test views can't be combined except as the start of
the chain.